requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
future==1.0.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
iso8601==2.1.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.0
mdurl==0.1.2
numpy==2.1.2
opencv-python==4.10.0.84
pydantic==2.9.2
pydantic_core==2.23.4
pydobot2==0.1.0
Pygments==2.18.0
pyserial==3.5
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
pyzbar==0.1.9
rich==13.9.2
serial==0.0.97
shellingham==1.5.4
sniffio==1.3.1
starlette==0.38.6
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
